requirements.txt
# Core Deep Learning Libraries
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Data Processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Evaluation Metrics
sacrebleu>=2.3.0
rouge-score>=0.1.2
nltk>=3.8.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.14.0

# Optional: Experiment Tracking
# wandb>=0.15.0

# Development Tools
jupyter>=1.0.0
ipywidgets>=8.0.0
